refactor(api): mv payment flat entity specific to chorus methods in chorus adapter
