+++ packages/api/src/modules/providers/chorus/adapters/ChorusRegionMapping.ts
@@ -1,0 +1,48 @@
+export const REGION_MAPPING = {
+    ADCE: "Administration Centrale",
+    DEFE: "Administration Centrale",
+    DILA: "Administration Centrale",
+    ETR1: "Administration Centrale",
+
+    DOM1: "DOM-TOM",
+    DOM2: "DOM-TOM",
+    DOM3: "DOM-TOM",
+    DOM4: "DOM-TOM",
+
+    ALSA: "Grand Est",
+    LORR: "Grand Est",
+    CHAR: "Grand Est",
+
+    AQUI: "Nouvelle-Aquitaine",
+    LIMO: "Nouvelle-Aquitaine",
+    POIT: "Nouvelle-Aquitaine",
+
+    AUVE: "Auvergne-Rhône-Alpes",
+    RALP: "Auvergne-Rhône-Alpes",
+
+    BNOR: "Normandie",
+    HNOR: "Normandie",
+
+    BOUR: "Bourgogne-Franche-Comté",
+    FRCO: "Bourgogne-Franche-Comté",
+
+    BRET: "Bretagne",
+
+    CENT: "Centre-Val de Loire",
+
+    CORS: "Corse",
+
+    IDF1: "Île-de-France",
+
+    LANG: "Occitanie",
+    MIPY: "Occitanie",
+
+    NORP: "Hauts-de-France",
+    PICA: "Hauts-de-France",
+
+    PACA: "Provence-Alpes-Côte d'Azur",
+
+    PAYL: "Pays de la Loire",
+};
+
+export default REGION_MAPPING;
